+++ app/src/main/java/org/linphone/activities/main/settings/viewmodels/AudioSettingsViewModel.kt
@@ -219,10 +219,10 @@ class AudioSettingsViewModel : GenericSettingsViewModel() {
         val default = core.defaultInputAudioDevice
 
         for (audioDevice in core.extendedAudioDevices) {
-            if (audioDevice.hasCapability(AudioDevice.Capabilities.CapabilityRecord)) {
+            if (audioDevice?.hasCapability(AudioDevice.Capabilities.CapabilityRecord) ?: false) {
                 labels.add(audioDevice.id)
                 values.add(audioDevice)
-                if (audioDevice.id == default.id) {
+                if (audioDevice.id == default?.id) {
                     inputAudioDeviceIndex.value = index
                 }
                 index += 1
@@ -242,7 +242,7 @@ class AudioSettingsViewModel : GenericSettingsViewModel() {
             if (audioDevice.hasCapability(AudioDevice.Capabilities.CapabilityPlay)) {
                 labels.add(audioDevice.id)
                 values.add(audioDevice)
-                if (audioDevice.id == default.id) {
+                if (audioDevice.id == default?.id) {
                     outputAudioDeviceIndex.value = index
                 }
                 index += 1