+++ app/src/main/java/org/linphone/model/callpopupjson.kt
@@ -1,0 +1,9 @@
+package org.linphone.model
+
+class callpopupjson(
+    var callerMsisdn: String,
+    var inTime: String,
+    var device_id: String,
+    var transactionId: String,
+    var callType: String
+)