+++ app/src/main/java/org/linphone/utils/DataBindingUtils.kt
@@ -25,10 +25,7 @@ import android.content.Context
 import android.text.Editable
 import android.text.TextWatcher
 import android.util.Patterns
-import android.view.LayoutInflater
-import android.view.TextureView
-import android.view.View
-import android.view.ViewGroup
+import android.view.*
 import android.view.inputmethod.EditorInfo
 import android.view.inputmethod.InputMethodManager
 import android.widget.*
@@ -228,6 +225,13 @@ fun setListener(view: SeekBar, lambda: (Any) -> Unit) {
     })
 }
 
+@BindingAdapter("inflatedLifecycleOwner")
+fun setInflatedViewStubLifecycleOwner(view: View, enable: Boolean) {
+    val binding = DataBindingUtil.bind<ViewDataBinding>(view)
+    // This is a bit hacky...
+    binding?.lifecycleOwner = view.context as GenericActivity
+}
+
 @BindingAdapter("entries")
 fun setEntries(
     viewGroup: ViewGroup,
@@ -337,11 +341,18 @@ private suspend fun loadContactPictureWithCoil(
     size: Int = 0,
     textSize: Int = 0,
     color: Int = 0,
-    textColor: Int = 0
+    textColor: Int = 0,
+    defaultAvatar: String? = null
 ) {
     val context = imageView.context
     if (contact == null) {
-        imageView.load(R.drawable.icon_single_contact_avatar)
+        if (defaultAvatar != null) {
+            imageView.load(defaultAvatar) {
+                transformations(CircleCropTransformation())
+            }
+        } else {
+            imageView.load(R.drawable.icon_single_contact_avatar)
+        }
     } else if (contact.showGroupChatAvatar) {
         imageView.load(AppCompatResources.getDrawable(context, R.drawable.icon_multiple_contacts_avatar))
     } else {
@@ -425,6 +436,21 @@ fun loadVoipContactPictureWithCoil(imageView: ImageView, contact: ContactDataInt
                 imageView, contact, false,
                 R.dimen.voip_contact_avatar_max_size, R.dimen.voip_contact_avatar_text_size,
                 R.attr.voipBackgroundColor, R.color.white_color
+            )
+        }
+    }
+}
+
+@BindingAdapter("coilSelfAvatar")
+fun loadSelfAvatarWithCoil(imageView: ImageView, contact: ContactDataInterface?) {
+    val coroutineScope = contact?.coroutineScope ?: coreContext.coroutineScope
+    coroutineScope.launch {
+        withContext(Dispatchers.Main) {
+            loadContactPictureWithCoil(
+                imageView, contact, false,
+                R.dimen.voip_contact_avatar_max_size, R.dimen.voip_contact_avatar_text_size,
+                R.attr.voipBackgroundColor, R.color.white_color,
+                corePreferences.defaultAccountAvatarPath
             )
         }
     }